Move some TC4820 functionality into a base class
